Move internal code out of storage/__init__.py
