+++ azure-storage/azure/storage/blobservice.py
@@ -39,33 +39,43 @@ from ._internal import (
     _ETreeXmlToObject,
     )
 from ._http import HTTPRequest
-from . import (
+from ._chunking import (
+    _BlockBlobChunkUploader,
+    _PageBlobChunkUploader,
+    _download_blob_chunks,
+    _upload_blob_chunks,
+)
+from .models import (
     Container,
     ContainerEnumResults,
     PageList,
     PageRange,
     SignedIdentifiers,
     StorageServiceProperties,
+)
+from .auth import (
     StorageSASAuthentication,
     StorageSharedKeyAuthentication,
     StorageNoAuthentication,
+)
+from .connection import (
     StorageConnectionParameters,
+)
+from .constants import (
     X_MS_VERSION,
-    _BlockBlobChunkUploader,
-    _PageBlobChunkUploader,
+)
+from ._serialization import (
     _convert_block_list_to_xml,
     _convert_response_to_block_list,
     _convert_signed_identifiers_to_xml,
     _create_blob_result,
-    _download_blob_chunks,
     _parse_blob_enum_results_list,
     _update_storage_blob_header,
-    _upload_blob_chunks,
-    )
+)
 from .sharedaccesssignature import (
     SharedAccessSignature,
     ResourceType,
-    )
+)
 from .storageclient import _StorageClient
 from os import path
 import sys
