+++ azure-storage/azure/storage/models.py
@@ -1,4 +1,4 @@
-#-------------------------------------------------------------------------
+﻿#-------------------------------------------------------------------------
 # Copyright (c) Microsoft.  All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,6 +14,7 @@
 #--------------------------------------------------------------------------
 from ._internal import (
     WindowsAzureData,
+    _list_of,
 )
